bigquery/sync-job/requirements.txt
pandas>=2.0.0
google-cloud-bigquery>=3.10.0
db-dtypes>=1.2.0
pyarrow>=14.0.0
supabase>=2.0.0
